refactor: 도메인과 상관없는 UI컴포넌트 src/ui로 이동
