+++ src/features/launch/private/components/LaunchPostContentControl.tsx
@@ -1,6 +1,6 @@
 import TextArea from '@/shared/ui/TextArea';
 import { useRecoilState } from 'recoil';
-import FormRowWide from '@/components/FormRowWide';
+import FormRowWide from '@/ui/FormRowWide';
 import { ChangeEvent } from 'react';
 import { postFormFieldSelector } from '@/features/launch/private/store/PostFormStateStore';
