+++ src/features/projectConfig/private/components/projectInfo/ConfigProjectDateControl.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import DateInput from '@/shared/ui/DateInput';
 import { useRecoilState } from 'recoil';
 import { addDays, format } from 'date-fns';
-import FormRow from '@/components/FormRow';
+import FormRow from '@/ui/FormRow';
 import { ProjectInfoSummary } from '@/service/project/public/getProjectPublicInfo';
 import { projectInfoFormSelector } from '@/features/projectConfig/private/store/ProjectInfoFormStateStore';
