+++ src/features/projectConfig/private/components/projectInfo/ConfigProjectNameControl.tsx
@@ -1,6 +1,6 @@
 import Input from '@/shared/ui/Input';
 import { useRecoilState } from 'recoil';
-import FormRow from '@/components/FormRow';
+import FormRow from '@/ui/FormRow';
 import { projectInfoFormSelector } from '@/features/projectConfig/private/store/ProjectInfoFormStateStore';
 import { ProjectInfoSummary } from '@/service/project/public/getProjectPublicInfo';
