+++ src/layouts/Header.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import logo from '../../public/images/logo.png';
 import Link from 'next/link';
-import LoginNav from '@/components/LoginNav';
+import LoginNav from '@/ui/LoginNav';
 import UserMenuSkeleton from '@/contents/user/private/UserMenuSkeleton';
 import { IoCreateOutline } from '@react-icons/all-files/io5/IoCreateOutline';
 import dynamic from 'next/dynamic';
